Added direction movement calculation. Changed position to double instead of int

+++ Server/src/main/java/com/dbs/dbs/models/units/Archer.java
@@ -5,7 +5,7 @@ package com.dbs.dbs.models.units;
  */
 public class Archer extends Unit{
 
-    public Archer(Integer posX, Integer posY) {
+    public Archer(Double posX, Double posY) {
         super("Archer", 40.0, 5.0, 6.0, 50.0, 75, posX, posY);
     }
 }
