+++ Server/src/main/java/com/dbs/dbs/models/units/Knight.java
@@ -5,7 +5,7 @@ package com.dbs.dbs.models.units;
  */
 public class Knight extends Unit{
 
-    public Knight(Integer posX, Integer posY) {
+    public Knight(Double posX, Double posY) {
         super("Knight", 7.0, 15.0, 20.0, 110.0, 150, posX, posY);
     }
 }
